+++ src/api/event-api.ts
@@ -1021,5 +1021,6 @@ export enum EventControllerGetConnectionsTypeEnum {
     attendee = 'attendee',
     admin = 'admin',
     owner = 'owner',
-    associated_group = 'associated_group'
+    associated_group = 'associated_group',
+    nearby = 'nearby'
 }