+++ src/models/mutate-group-entity-connection-dto.ts
@@ -24,7 +24,8 @@ export interface MutateGroupEntityConnectionDto {
 export enum MutateGroupEntityConnectionDtoTypeEnum {
     member = 'member',
     admin = 'admin',
-    owner = 'owner'
+    owner = 'owner',
+    nearby = 'nearby'
 }