+++ src/models/user-entity-connection-dto.ts
@@ -42,6 +42,7 @@ export enum UserEntityConnectionDtoTypeEnum {
     admin = 'admin',
     owner = 'owner',
     associated_group = 'associated_group',
+    nearby = 'nearby',
     member = 'member',
     friend = 'friend',
     blocked = 'blocked',